python_server/requirements.txt
Flask==3.0.0
tensorflow==2.21.0
Pillow==10.2.0
numpy==1.26.4
